korekcije kod izlaznica automatska odjava

+++ database/migrations/2025_07_04_000000_create_redosljed_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('redosljed', function (Blueprint $table) {
             $table->id(); // auto increment primary key
             $table->string('radno_mjesto');
-            $table->integer('redni_broj');
+            $table->float('redni_broj');
             $table->timestamps();
         });
     }
